test(sveltekit): Enable native OTel tracing in SvelteKit 3 e2e app
Turn on SvelteKit 3's native server-side tracing (`experimental.tracing.server`
and `experimental.instrumentation.server`) so the e2e app exercises the path
where @sentry/sveltekit picks up Kit's OpenTelemetry spans instead of starting
its own `http.server` span.

`Sentry.init` moves to `src/instrumentation.server.ts` (the native instrumentation
entry point); `hooks.server.ts` is reduced to `sequence(sentryHandle())` plus the
error handler. Swap the default-path performance/sdk tests for the native-tracing
`tracing.*` suites and add a `form-action` route to cover form-action spans,
mirroring `sveltekit-2-kit-tracing`. Config is passed via the `sveltekit()` Vite
plugin since SvelteKit 3 dropped `svelte.config.js`.
