+++ dev-packages/e2e-tests/test-applications/sveltekit-3/src/hooks.server.ts
@@ -1,15 +1,7 @@
-import { E2E_TEST_DSN } from '$env/static/private';
 import * as Sentry from '@sentry/sveltekit';
-
-Sentry.init({
-  environment: 'qa', // dynamic sampling bias to keep transactions
-  dsn: E2E_TEST_DSN,
-  debug: !!process.env.DEBUG,
-  tunnel: `http://localhost:3031/`, // proxy server
-  tracesSampleRate: 1.0,
-});
+import { sequence } from '@sveltejs/kit/hooks';
 
 // not logging anything to console to avoid noise in the test output
 export const handleError = Sentry.handleErrorWithSentry(() => {});
 
-export const handle = Sentry.sentryHandle();
+export const handle = sequence(Sentry.sentryHandle());
