+++ dev-packages/e2e-tests/test-applications/sveltekit-3/src/routes/form-action/+page.server.ts
@@ -1,0 +1,7 @@
+export const actions = {
+  default: async ({ request }) => {
+    const formData = await request.formData();
+    const name = formData.get('name');
+    return { name };
+  },
+};
