+++ dev-packages/e2e-tests/test-applications/sveltekit-3/src/routes/form-action/+page.svelte
@@ -1,0 +1,15 @@
+<script>
+  import { enhance } from '$app/forms';
+
+  const { data, form } = $props();
+</script>
+
+<form method="post" use:enhance>
+  <label for="name">Name:</label>
+  <input id="inputName" type="text" name="name" />
+  <button id="buttonSubmit" type="submit">Submit</button>
+</form>
+
+{#if form?.name}
+  <p>Hello {form.name}</p>
+{/if}
