+++ dev-packages/e2e-tests/test-applications/sveltekit-3/tests/tracing.client.test.ts
@@ -4,23 +4,28 @@ import { waitForInitialPageload } from './utils';
 
 test.describe('client-specific performance events', () => {
   test('multiple navigations have distinct traces', async ({ page }) => {
-    const navigationTxn1EventPromise = waitForTransaction('sveltekit-2', txnEvent => {
+    const navigationTxn1EventPromise = waitForTransaction('sveltekit-3', txnEvent => {
       return txnEvent?.transaction === '/nav1' && txnEvent.contexts?.trace?.op === 'navigation';
     });
 
-    const navigationTxn2EventPromise = waitForTransaction('sveltekit-2', txnEvent => {
+    const navigationTxn2EventPromise = waitForTransaction('sveltekit-3', txnEvent => {
       return txnEvent?.transaction === '/' && txnEvent.contexts?.trace?.op === 'navigation';
     });
 
-    const navigationTxn3EventPromise = waitForTransaction('sveltekit-2', txnEvent => {
+    const navigationTxn3EventPromise = waitForTransaction('sveltekit-3', txnEvent => {
       return txnEvent?.transaction === '/nav2' && txnEvent.contexts?.trace?.op === 'navigation';
     });
 
     await waitForInitialPageload(page);
 
-    const [navigationTxn1Event] = await Promise.all([navigationTxn1EventPromise, page.getByText('Nav 1').click()]);
-    const [navigationTxn2Event] = await Promise.all([navigationTxn2EventPromise, page.goBack()]);
-    const [navigationTxn3Event] = await Promise.all([navigationTxn3EventPromise, page.getByText('Nav 2').click()]);
+    await page.getByText('Nav 1').click();
+    const navigationTxn1Event = await navigationTxn1EventPromise;
+
+    await page.goBack();
+    const navigationTxn2Event = await navigationTxn2EventPromise;
+
+    await page.getByText('Nav 2').click();
+    const navigationTxn3Event = await navigationTxn3EventPromise;
 
     expect(navigationTxn1Event).toMatchObject({
       transaction: '/nav1',
@@ -68,7 +73,7 @@ test.describe('client-specific performance events', () => {
   });
 
   test('records manually added component tracking spans', async ({ page }) => {
-    const componentTxnEventPromise = waitForTransaction('sveltekit-2', txnEvent => {
+    const componentTxnEventPromise = waitForTransaction('sveltekit-3', txnEvent => {
       return txnEvent?.transaction === '/components';
     });
 
@@ -102,30 +107,6 @@ test.describe('client-specific performance events', () => {
           data: { 'sentry.op': 'ui.svelte.init', 'sentry.origin': 'auto.ui.svelte' },
           description: '<Component3>',
           op: 'ui.svelte.init',
-          origin: 'auto.ui.svelte',
-        }),
-        expect.objectContaining({
-          data: { 'sentry.op': 'ui.svelte.update', 'sentry.origin': 'auto.ui.svelte' },
-          description: '<components/+page>',
-          op: 'ui.svelte.update',
-          origin: 'auto.ui.svelte',
-        }),
-        expect.objectContaining({
-          data: { 'sentry.op': 'ui.svelte.update', 'sentry.origin': 'auto.ui.svelte' },
-          description: '<Component1>',
-          op: 'ui.svelte.update',
-          origin: 'auto.ui.svelte',
-        }),
-        expect.objectContaining({
-          data: { 'sentry.op': 'ui.svelte.update', 'sentry.origin': 'auto.ui.svelte' },
-          description: '<Component2>',
-          op: 'ui.svelte.update',
-          origin: 'auto.ui.svelte',
-        }),
-        expect.objectContaining({
-          data: { 'sentry.op': 'ui.svelte.update', 'sentry.origin': 'auto.ui.svelte' },
-          description: '<Component3>',
-          op: 'ui.svelte.update',
           origin: 'auto.ui.svelte',
         }),
       ]),
