+++ dev-packages/e2e-tests/test-applications/sveltekit-3/tests/utils.ts
@@ -24,7 +24,7 @@ export async function waitForInitialPageload(
   const txnName = opts?.parameterizedRoute ?? route;
   const debug = opts?.debug ?? false;
 
-  const clientPageloadTxnEventPromise = waitForTransaction('sveltekit-2', txnEvent => {
+  const clientPageloadTxnEventPromise = waitForTransaction('sveltekit-3', txnEvent => {
     debug &&
       console.log({
         txn: txnEvent?.transaction,
